Each app owns its sidebar; shared chassis moves to platform
SidebarFrame (platform/ui/sidebar) keeps the shared chrome — brand row,
draggable width, collapse strip, resize handle — while every owner
composes its own body: ShellSidebar (app-switcher + settings),
ExplorerSidebar (Home + bookmarks + file recents), BuilderSidebar
(Home + app recents). App.tsx picks the component by route instead of
threading a ctx flag; the SidebarCtx type is gone.
